Move Firebase client factory to client impl.
